requirements.txt
streamlit
sentence-transformers
transformers
chromadb>=0.4.0
gitpython
requests